Smarty: Enhanced includes
As includes are now compiled pre-emptively, we had a problem with infinite regression if the include then included itself.  To get around this, includes are now compiled as closures and then indexed.  If the include already exists, then it doesn't recompile it.  Then an include will just execute the compiled closure and we are good to go!