More tweaks to the controller router
This is a little bit more robust in finding controllers, but also makes the assumption that if a controller sub directory exists, but processing has stopped, default to the Index controller for that sub dir.  If it doesn't exist, then we get a standard 404 later when the controller is loaded.  By making this assumption it's now possible to have a default controller in a subdirectory WITH a default action, meaning that any requests that do not find a specific controller will default to the index controller and default action as a global handler.  Much coolness.