Auth session cookie now only exists if session is active/used
We delay setting the cookie until a Cache::set() is called.  Which means if it is never called (because the user never authenticates) the cookie is never set.  Also, then once deauth occurs the cookie is removed.  This can increase security by not having unnecessary cookie flying around the wire.