Updated how 'base' config attribute works
Fortunately this is a little known config parameter and is only used in our personal server configs.  This allows the base URL to be set specifically.  This works great for a base that has no root path.  But what I have done is update this to override the root path as previously this only overrode the host part.