Updated dataBinderArray watchers
This breaks old functionality, but is now more consistent with the dataBinder.watch() function in that the callback arguments are almost the same.  The first argument is the new item.  In the case of a list, the second argument will now be the new item clones from the list template.